+++ packages/solid-crs-core/lib/loan/models/loan-request.ts
@@ -4,6 +4,10 @@ export interface LoanRequest {
    */
   uri: string;
   /**
+   *
+   */
+  type: 'https://www.w3.org/ns/activitystreams#Accept' | 'https://www.w3.org/ns/activitystreams#Reject' | 'https://www.w3.org/ns/activitystreams#Offer';
+  /**
    * The uri of the requested collection
    */
   collection: string;
@@ -23,9 +27,4 @@ export interface LoanRequest {
    * An optional note for the receiving institution
    */
   description?: string;
-  /**
-   * Indicates if the request is accepted or not
-   * Optional: when this parameter is undefined, the request is pending / neither accepted or rejected
-   */
-  accepted?: boolean;
 }