+++ packages/solid-crs-manage/lib/features/loan/loan.machine.ts
@@ -1,8 +1,8 @@
 import { assign, DoneInvokeEvent, MachineConfig } from 'xstate';
 import { log } from 'xstate/lib/actions';
-import { LoanRequest } from '@netwerk-digitaal-erfgoed/solid-crs-core';
+import { LoanRequest, Collection } from '@netwerk-digitaal-erfgoed/solid-crs-core';
 import { LoanContext } from './loan.context';
-import { ClickedAcceptedLoanRequestEvent, ClickedRejectedLoanRequestEvent, ClickedSendLoanRequestEvent, LoanEvent, LoanEvents } from './loan.events';
+import { ClickedLoanRequestDetailEvent, LoanEvent, LoanEvents } from './loan.events';
 import * as services from './loan.services';
 import { LoanStates, LoanStateSchema } from './loan.states';
 
@@ -11,24 +11,54 @@ export const loanMachine: MachineConfig<LoanContext, LoanStateSchema, LoanEvent>
   states: {
     [LoanStates.LOADING_LOAN_REQUESTS]: {
       invoke: {
-        src: () => services.loadRequests(),
+        src: (c, e) => services.loadRequests(c, e),
         onDone: {
-          actions: assign({ loanRequests: (context, event: DoneInvokeEvent<LoanRequest[]>) => event.data }),
-          target: LoanStates.LOAN_REQUEST_OVERVIEW,
+          actions: assign({ loanRequests: (c, event: DoneInvokeEvent<LoanRequest[]>) => event.data }),
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_INCOMING,
         },
         onError: {
-          actions: log('Error Loading Requests'),
+          actions: log((c, e) => `Error Loading Request: ${e.data}`),
         },
       },
     },
-    [LoanStates.LOAN_REQUEST_OVERVIEW]: {
+    [LoanStates.LOADING_COLLECTION]: {
+      invoke: {
+        src: (context, event: ClickedLoanRequestDetailEvent) =>
+          context.collectionStore.get(event.loanRequest.collection),
+        onDone: {
+          actions: assign({ collection: (c, event: DoneInvokeEvent<Collection>) => event.data }),
+          target: LoanStates.LOAN_REQUEST_DETAIL,
+        },
+        onError: {
+          actions: log((c, e) => `Error Loading Collection: ${e.data}`),
+        },
+      },
+    },
+    [LoanStates.LOAN_REQUEST_OVERVIEW_INCOMING]: {
       on: {
+        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW_ACCEPTED]: {
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_ACCEPTED,
+        },
         [LoanEvents.CLICKED_NEW_LOAN_REQUEST]: {
           target: LoanStates.LOAN_REQUEST_CREATION,
         },
         [LoanEvents.CLICKED_LOAN_REQUEST_DETAIL]: {
-          actions: assign({ loanRequest: (context, event) => event.loanRequest }),
-          target: LoanStates.LOAN_REQUEST_DETAIL,
+          actions: assign({ loanRequest: (c, event) => event.loanRequest }),
+          target: LoanStates.LOADING_COLLECTION,
+        },
+      },
+    },
+    [LoanStates.LOAN_REQUEST_OVERVIEW_ACCEPTED]: {
+      on: {
+        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW_INCOMING]: {
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_INCOMING,
+        },
+        [LoanEvents.CLICKED_NEW_LOAN_REQUEST]: {
+          target: LoanStates.LOAN_REQUEST_CREATION,
+        },
+        [LoanEvents.CLICKED_LOAN_REQUEST_DETAIL]: {
+          actions: assign({ loanRequest: (c, event) => event.loanRequest }),
+          target: LoanStates.LOADING_COLLECTION,
         },
       },
     },
@@ -37,8 +67,11 @@ export const loanMachine: MachineConfig<LoanContext, LoanStateSchema, LoanEvent>
         [LoanEvents.CLICKED_NEW_LOAN_REQUEST]: {
           target: LoanStates.LOAN_REQUEST_CREATION,
         },
-        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW]: {
-          target: LoanStates.LOAN_REQUEST_OVERVIEW,
+        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW_INCOMING]: {
+          target: LoanStates.LOADING_LOAN_REQUESTS,
+        },
+        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW_ACCEPTED]: {
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_ACCEPTED,
         },
         [LoanEvents.CLICKED_ACCEPTED_LOAN_REQUEST]: {
           target: LoanStates.ACCEPTING_LOAN_REQUEST,
@@ -50,8 +83,11 @@ export const loanMachine: MachineConfig<LoanContext, LoanStateSchema, LoanEvent>
     },
     [LoanStates.LOAN_REQUEST_CREATION]: {
       on: {
-        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW]: {
-          target: LoanStates.LOAN_REQUEST_OVERVIEW,
+        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW_INCOMING]: {
+          target: LoanStates.LOADING_LOAN_REQUESTS,
+        },
+        [LoanEvents.CLICKED_LOAN_REQUEST_OVERVIEW_ACCEPTED]: {
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_ACCEPTED,
         },
         [LoanEvents.CLICKED_SEND_LOAN_REQUEST]: {
           target: LoanStates.SENDING_LOAN_REQUEST,
@@ -60,34 +96,34 @@ export const loanMachine: MachineConfig<LoanContext, LoanStateSchema, LoanEvent>
     },
     [LoanStates.SENDING_LOAN_REQUEST]: {
       invoke: {
-        src: (context, event: ClickedSendLoanRequestEvent) => services.createRequest(event.loanRequestCreationArgs),
+        src: (c, e) => services.createRequest(c, e),
         onDone: {
-          target: LoanStates.LOAN_REQUEST_OVERVIEW,
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_INCOMING,
         },
         onError: {
-          actions: log('Error Creating Request'),
+          actions: log((c, e) => `Error Creating Request: ${e.data}`),
         },
       },
     },
     [LoanStates.ACCEPTING_LOAN_REQUEST]: {
       invoke: {
-        src: (context, event: ClickedAcceptedLoanRequestEvent) => services.acceptRequest(event.loanRequest),
+        src: (c, e) => services.acceptRequest(c, e),
         onDone: {
-          target: LoanStates.LOAN_REQUEST_OVERVIEW,
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_INCOMING,
         },
         onError: {
-          actions: log('Error Accepting Request'),
+          actions: log((c, e) => `Error Accepting Request: ${e.data}`),
         },
       },
     },
     [LoanStates.REJECTING_LOAN_REQUEST]: {
       invoke: {
-        src: (context, event: ClickedRejectedLoanRequestEvent) => services.rejectRequest(event.loanRequest),
+        src: (c, e) => services.rejectRequest(c, e),
         onDone: {
-          target: LoanStates.LOAN_REQUEST_OVERVIEW,
+          target: LoanStates.LOAN_REQUEST_OVERVIEW_INCOMING,
         },
         onError: {
-          actions: log('Error Rejecting Request'),
+          actions: log((c, e) => `Error Rejecting Request: ${e.data}`),
         },
       },
     },