+++ packages/solid-crs-client/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.0-loans.19](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.19) (2022-03-15)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-client
+
+
+
+
+
+## [0.17.0-loans.18](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.18) (2022-03-15)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-client
+
+
+
+
+
 ## [0.17.0-loans.17](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.17) (2022-03-09)
 
 **Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-client