+++ packages/solid-crs-manage/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.0-loans.19](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.19) (2022-03-15)
+
+
+### **Bug Fixes**
+
+* package-locks for loans ([#586](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/586)) ([fe90a07](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/fe90a072ca8bf61da35efe671d22fc8f5d13d51b))
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* add object updates page in object root ([#577](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/577)) ([a733f99](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a733f996c3fbff5e2a6df690f223357420dd6d15))
+* create loan feature components ([#568](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/568)) ([6b77955](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/6b77955da099266ab7f10c3e3920e8f23451a6e0)), closes [#559](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/559)
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+* implement LDNs for loan requests ([#573](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/573)) ([90b25b9](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/90b25b96c21d86c9be8e494dbc60b3ad85c77418)), closes [#575](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/575)
+* loan feature setup ([#564](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/564)) ([674e27c](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/674e27cc838f442e4e8c16444f7a6ba17d7f0bca)), closes [#559](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/559)
+
+
+
+## [0.17.0-loans.18](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.18) (2022-03-15)
+
+
+### **Bug Fixes**
+
+* package-locks for loans ([#586](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/586)) ([fe90a07](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/fe90a072ca8bf61da35efe671d22fc8f5d13d51b))
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* add object updates page in object root ([#577](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/577)) ([a733f99](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a733f996c3fbff5e2a6df690f223357420dd6d15))
+* create loan feature components ([#568](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/568)) ([6b77955](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/6b77955da099266ab7f10c3e3920e8f23451a6e0)), closes [#559](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/559)
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+* loan feature setup ([#564](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/564)) ([674e27c](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/674e27cc838f442e4e8c16444f7a6ba17d7f0bca)), closes [#559](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/559)
+
+
+
 ## [0.17.0-loans.17](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.17) (2022-03-09)